+++ packages/common/src/utils/signature_utils.ts
@@ -4,6 +4,7 @@ import type { CommitData } from '../index.js'
 import type { StreamID } from '@ceramicnetwork/streamid'
 import { getEIP191Verifier } from '@didtools/pkh-ethereum'
 import { getSolanaVerifier } from '@didtools/pkh-solana'
+import { getStacksVerifier } from '@didtools/pkh-stacks'
 import { getTezosVerifier } from '@didtools/pkh-tezos'
 
 const DEFAULT_CACAO_REVOCATION_PHASE_OUT = 24 * 60 * 60
@@ -12,6 +13,7 @@ const DEFAULT_CACAO_REVOCATION_PHASE_OUT = 24 * 60 * 60
 const verifiersCACAO = {
   ...getEIP191Verifier(),
   ...getSolanaVerifier(),
+  ...getStacksVerifier(),
   ...getTezosVerifier(),
 }