+++ packages/common/src/utils/stream-utils.ts
@@ -12,11 +12,12 @@ import {
   SignedCommit,
   SignedCommitContainer,
 } from '../index.js'
-import { AnchorStatus, LogEntry, StreamState } from '../stream.js'
+import { LogEntry, StreamState } from '../stream.js'
 import type { DagJWS } from 'dids'
 import { StreamID, StreamType } from '@ceramicnetwork/streamid'
 import { CID } from 'multiformats/cid'
 import { base64urlToJSON } from '@ceramicnetwork/common'
+import { AnchorStatus } from '@ceramicnetwork/codecs'
 
 const TILE_TYPE_ID = 0