+++ packages/stream-tests/src/__tests__/capability.test.ts
@@ -1,4 +1,11 @@
-import { AnchorStatus, CeramicApi, IpfsApi, SyncOptions, TestUtils } from '@ceramicnetwork/common'
+import {
+  AnchorStatus,
+  CeramicApi,
+  IpfsApi,
+  StreamUtils,
+  SyncOptions,
+  TestUtils,
+} from '@ceramicnetwork/common'
 import { createIPFS } from '@ceramicnetwork/ipfs-daemon'
 import { TileDocument } from '@ceramicnetwork/stream-tile'
 import { DID } from 'dids'