+++ packages/stream-tests/src/__tests__/cross-node-indexing.test.ts
@@ -1,6 +1,5 @@
 import { jest } from '@jest/globals'
 import {
-  CeramicApi,
   IpfsApi,
   Page,
   PaginationQuery,