+++ packages/anchor-utils/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.9.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-utils@1.8.0...@ceramicnetwork/anchor-utils@1.9.0-rc.0) (2023-05-24)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-utils
+
+
+
+
+
+# [1.8.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-utils@1.8.0-rc.0...@ceramicnetwork/anchor-utils@1.8.0) (2023-05-03)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-utils
+
+
+
+
+
+# [1.8.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-utils@1.7.0...@ceramicnetwork/anchor-utils@1.8.0-rc.0) (2023-04-26)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-utils
+
+
+
+
+
+# [1.7.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-utils@1.7.0-rc.0...@ceramicnetwork/anchor-utils@1.7.0) (2023-03-22)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-utils
+
+
+
+
+
+# [1.7.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-utils@1.6.0...@ceramicnetwork/anchor-utils@1.7.0-rc.0) (2023-03-16)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-utils
+
+
+
+
+
 # [1.6.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-utils@1.6.0-rc.0...@ceramicnetwork/anchor-utils@1.6.0) (2023-02-28)
 
 **Note:** Version bump only for package @ceramicnetwork/anchor-utils